refactoring because i hate myself
beneficiaries work ok, idk others yes
