--- sample-data/unit testing ;3c/beneficiaries-unit testing.csv
@@ -1,24 +1,0 @@
-Child Number (ID)	First Name	Last Name	Sex	Birthdate	Grade Level	Address	Name (Guardian 1)	Relation (Guardian 1)	Contact Number (Guardian 1)	Email (Guardian 1)	Name (Guardian 2)	Relation (Guardian 2)	Contact Number (Guardian 2)	Email (Guardian 2)	Name (Guardian 3)	Relation (Guardian 3)	Contact Number (Guardian 3)	Email (Guardian 3)
-StringId	String	Id	M	11/12/2004	1	should skip	test	test	09172647328	test@gmail.com								
-1		No FN	M	11/12/2004	1	should skip	test	test	09172647328	test@gmail.com								
-2	No LN		M	11/12/2004	1	should skip	test	test	09172647328	test@gmail.com								
-3	Blank	Sex		11/12/2004	1	should keep	test	test	09172647328	test@gmail.com								
-4	Invalid	Sex	Suzie Toot	11/12/2004	1	should skip	test	test	09172647328	test@gmail.com								
-5	Blank	Date	M		1	should keep	test	test	09172647328	test@gmail.com								
-6	Invalid1	Date	M	3232233	1	should skip	test	test	09172647328	test@gmail.com								
-7	Invalid2	Date	M	datewoww	1	should skip	test	test	09172647328	test@gmail.com								
-8	Blank	Grade	M	11/12/2004		should keep	test	test	09172647328	test@gmail.com								
-9	Invalid1	Grade	M	11/12/2004	34	should skip	test	test	09172647328	test@gmail.com								
-10	Invalid2	Grade	M	11/12/2004	Grades	should skip	test	test	09172647328	test@gmail.com								
-11	Blank	Address	M	11/12/2004	1		keep	test	09172647328	test@gmail.com								
-12	blank	guardian	M	11/12/2004	1	should keep	  											
-13	no Contact	guardian	M	11/12/2004	1	keep?	test	test										
-14	no Name Rel	guardian	M	11/12/2004	1	keep			09172647328	test@gmail.com								
-15	invalid contact	guardian	M	11/12/2004	1	should skip	test	test	92039203929	test@gmail.com								
-16	invalid email	guardian	M	11/12/2004	1	keep	test	test	09172647328	my email								
-17	blank	guardian	M	11/12/2004	1	should keep												
-18	first blank	guardian	M	11/12/2004	1	should keep					test	test	09172647328	test@gmail.com	test	test	09172647328	test@gmail.com
-19	2nd blank	guardian	M	11/12/2004	1	should keep	test	test	09172647328	test@gmail.com					test	test	09172647328	test@gmail.com
-20	2blank	guardian	M	11/12/2004	1	should keep									test	test	09172647328	test@gmail.com
-21	Normal	Benef	M	11/12/2004	1	should keep	test	test	09172647328	test@gmail.com								
-21	Dupe	Id	M	11/12/2004	1	should skip	test	test	09172647328	test@gmail.com
